Bump Numpy to 2.0.0. Bump AllanTools to 2024.6
This change slightly modifies the results. This is insignificant as it is the result of a different rounding due to the updated floating point libraries. This is most visible when small deviations like the temperature
coefficients are calculated and can be observed for the least significant digits. These minor deviations do not change the final results in any way, but can be detected in the plots.